refactor: encapsulate the policy attachment logic
