+++ pkg/gateway/cache/faultinjection_trigger.go
@@ -6,12 +6,12 @@ import (
 	"github.com/flomesh-io/fsm/pkg/gateway/utils"
 )
 
-// FaultInjectionPoliciesProcessor is responsible for processing FaultInjectionPolicy objects
-type FaultInjectionPoliciesProcessor struct {
+// FaultInjectionPoliciesTrigger is responsible for processing FaultInjectionPolicy objects
+type FaultInjectionPoliciesTrigger struct {
 }
 
 // Insert adds a FaultInjectionPolicy to the cache and returns true if the target service is routable
-func (p *FaultInjectionPoliciesProcessor) Insert(obj interface{}, cache *GatewayCache) bool {
+func (p *FaultInjectionPoliciesTrigger) Insert(obj interface{}, cache *GatewayCache) bool {
 	policy, ok := obj.(*gwpav1alpha1.FaultInjectionPolicy)
 	if !ok {
 		log.Error().Msgf("unexpected object type %T", obj)
@@ -24,7 +24,7 @@ func (p *FaultInjectionPoliciesProcessor) Insert(obj interface{}, cache *Gateway
 }
 
 // Delete removes a FaultInjectionPolicy from the cache and returns true if the policy was found
-func (p *FaultInjectionPoliciesProcessor) Delete(obj interface{}, cache *GatewayCache) bool {
+func (p *FaultInjectionPoliciesTrigger) Delete(obj interface{}, cache *GatewayCache) bool {
 	policy, ok := obj.(*gwpav1alpha1.FaultInjectionPolicy)
 	if !ok {
 		log.Error().Msgf("unexpected object type %T", obj)
