+++ pkg/gateway/cache/httproutes_trigger.go
@@ -6,12 +6,12 @@ import (
 	"github.com/flomesh-io/fsm/pkg/gateway/utils"
 )
 
-// HTTPRoutesProcessor is responsible for processing HTTPRoute objects
-type HTTPRoutesProcessor struct {
+// HTTPRoutesTrigger is responsible for processing HTTPRoute objects
+type HTTPRoutesTrigger struct {
 }
 
 // Insert adds a HTTPRoute to the cache and returns true if the route is effective
-func (p *HTTPRoutesProcessor) Insert(obj interface{}, cache *GatewayCache) bool {
+func (p *HTTPRoutesTrigger) Insert(obj interface{}, cache *GatewayCache) bool {
 	route, ok := obj.(*gwv1beta1.HTTPRoute)
 	if !ok {
 		log.Error().Msgf("unexpected object type %T", obj)
@@ -24,7 +24,7 @@ func (p *HTTPRoutesProcessor) Insert(obj interface{}, cache *GatewayCache) bool 
 }
 
 // Delete removes a HTTPRoute from the cache and returns true if the route was found
-func (p *HTTPRoutesProcessor) Delete(obj interface{}, cache *GatewayCache) bool {
+func (p *HTTPRoutesTrigger) Delete(obj interface{}, cache *GatewayCache) bool {
 	route, ok := obj.(*gwv1beta1.HTTPRoute)
 	if !ok {
 		log.Error().Msgf("unexpected object type %T", obj)
