+++ pkg/gateway/cache/service_trigger.go
@@ -6,11 +6,11 @@ import (
 	"github.com/flomesh-io/fsm/pkg/gateway/utils"
 )
 
-// ServicesProcessor is responsible for processing Service objects
-type ServicesProcessor struct{}
+// ServicesTrigger is responsible for processing Service objects
+type ServicesTrigger struct{}
 
 // Insert adds the Service object to the cache and returns true if the cache was modified
-func (p *ServicesProcessor) Insert(obj interface{}, cache *GatewayCache) bool {
+func (p *ServicesTrigger) Insert(obj interface{}, cache *GatewayCache) bool {
 	svc, ok := obj.(*corev1.Service)
 	if !ok {
 		log.Error().Msgf("unexpected object type %T", obj)
@@ -24,7 +24,7 @@ func (p *ServicesProcessor) Insert(obj interface{}, cache *GatewayCache) bool {
 }
 
 // Delete removes the Service object from the cache and returns true if the cache was modified
-func (p *ServicesProcessor) Delete(obj interface{}, cache *GatewayCache) bool {
+func (p *ServicesTrigger) Delete(obj interface{}, cache *GatewayCache) bool {
 	svc, ok := obj.(*corev1.Service)
 	if !ok {
 		log.Error().Msgf("unexpected object type %T", obj)
