+++ pkg/gateway/cache/tlsroutes_trigger.go
@@ -6,12 +6,12 @@ import (
 	"github.com/flomesh-io/fsm/pkg/gateway/utils"
 )
 
-// TLSRoutesProcessor is responsible for processing TLSRoute objects
-type TLSRoutesProcessor struct {
+// TLSRoutesTrigger is responsible for processing TLSRoute objects
+type TLSRoutesTrigger struct {
 }
 
 // Insert adds a TLSRoute to the cache and returns true if the route is effective
-func (p *TLSRoutesProcessor) Insert(obj interface{}, cache *GatewayCache) bool {
+func (p *TLSRoutesTrigger) Insert(obj interface{}, cache *GatewayCache) bool {
 	route, ok := obj.(*gwv1alpha2.TLSRoute)
 	if !ok {
 		log.Error().Msgf("unexpected object type %T", obj)
@@ -24,7 +24,7 @@ func (p *TLSRoutesProcessor) Insert(obj interface{}, cache *GatewayCache) bool {
 }
 
 // Delete removes a TLSRoute from the cache and returns true if the route was found
-func (p *TLSRoutesProcessor) Delete(obj interface{}, cache *GatewayCache) bool {
+func (p *TLSRoutesTrigger) Delete(obj interface{}, cache *GatewayCache) bool {
 	route, ok := obj.(*gwv1alpha2.TLSRoute)
 	if !ok {
 		log.Error().Msgf("unexpected object type %T", obj)
